Prevent 'normal' user from using db-clear

+++ droptables.php
@@ -1,12 +1,22 @@
-<?php 
+<?php
 ob_start();
 require './config/config.php';
-include'frontend/functions.php';
+include 'frontend/functions.php';
 include("login/auth.php");
 
-if(isset($_SESSION["uname"])){ 
-$sql = "DROP TABLE `gyms`, `pokedex`, `quests`, `raidbosses`, `rewards`, `spotraid`, `spots`, `stops`, `teams`, `usergroup`, `users`"; 
-mysqli_query($conn,$sql);
+if (isset($_SESSION["uname"])) {
+    
+    $result = $conn->query("SELECT * FROM users,usergroup WHERE uname='" . $_SESSION['uname'] . "' AND users.usergroup = usergroup.id LIMIT 1  ");
+    $id     = $usergroup = "";
+    while ($row = $result->fetch_assoc()) {
+        $id        = $row['id'];
+        $uname     = $row['uname'];
+        $email     = $row['email'];
+        $usergroup = $row['groupname'];
+        if ("$usergroup" == 'admin') {
+            $sql = "DROP TABLE `gyms`, `pokedex`, `quests`, `raidbosses`, `rewards`, `spotraid`, `spots`, `stops`, `teams`, `usergroup`, `users`";
+            mysqli_query($conn, $sql);
+            
 ?>
 <html>
 <head>
@@ -28,7 +38,7 @@ mysqli_query($conn,$sql);
 </body>
 </html>
 <?php
-} else {
+        } else {
 ?>
 <html>
 <head>
@@ -50,7 +60,29 @@ Click here to <a href='./login/login.php'>Login</a>
 </body>
 </html>
 <?php
+        }
+    }
+} else {
+?>
+<html>
+<head>
+<meta charset="utf-8">
+<title>Not logged in</title>
+<meta content="True" name="HandheldFriendly">
+<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0">
+<meta name="viewport" content="width=device-width">
+<link rel="stylesheet" href="./login/css/style.css" />
+</head>
+<body>
+<center>
+<div class='form'>
+<h1>Error</h1>
+<h3>Please login to perform this action</h3><br>
+Click here to <a href='./login/login.php'>Login</a>
+</div>
+</center>
+</body>
+</html>
+<?php
 }
 ?>
-
-
